requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
numpy==1.26.2
Pillow==10.1.0
librosa==0.10.1
matplotlib==3.8.2
moviepy==1.0.3
soundfile==0.12.1

